added support for forgery

+++ blog/blog_api/api.py
@@ -4,8 +4,15 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from serializers import ArticleLikesSerializer
 from blog.blog import ArticleLikes
-
+from rest_framework.exceptions import APIException
 from rest_framework.views import exception_handler
+
+
+
+class NotAllowedError(APIException):
+    status_code = 504
+    default_detail = 'your are not allowed to do the work of others.'
+    default_code = 'fraudulant '
 
 def rest_exc_handler(exc, context):
     # Call REST framework's default exception handler first,
@@ -44,17 +51,28 @@ def article_likes(request, pk):
         return Response(serializer.data)
 
     elif request.method == 'POST':
-        serializer = ArticleLikesSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        import pdb
+        pdb.set_trace()
+        if request.data.get('user_id') == request.user.id:
+            serializer = ArticleLikesSerializer(data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                serializer.data.update({'total_likes':len(article_likes)})
+                import pdb
+                pdb.set_trace()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            # return Response(status=status.HTTP_400_BAD_REQUEST)
+            raise NotAllowedError
     
     elif request.method == 'DELETE':
-        try:
-            like = ArticleLikes.objects.filter(article_id=pk, user_id=request.user)
-            like.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except :
-            return Response(status=status.HTTP_404_NOT_FOUND)
-            
+        if request.data.get('user_id') == request.user.id:
+            try:
+                like = ArticleLikes.objects.filter(article_id=pk, user_id=request.user)
+                like.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            except :
+                return Response(status=status.HTTP_404_NOT_FOUND)
+        else:
+            raise NotAllowedError()
